Speedup user list page loading time
Fetching a lot of data in the "getInitialProps", will cause the server
to wait for the data before responding to the browser. By moving the API
calls from "getInitialProps" into the "componentDidMount", the server
will respond faster. After the components are mounted in the browser,
"componentDidMount" will be invoked and start to call the API.