Remove showClaims action
Redux = Action + Reducer
Redux holds the state in the one place.
When action is dispatched, the reducer will be called and change the
Redux state.

Redux is suitable to hold the data that need to be shared between pages,
such as user profile and session token.

Since the claim data only used in the ShowClaimPage page, the claim data
is not necessary to behold in the Redux state, so we remove 'ShowClaims'
action and related codes.