Preserve default environment values
Since we are using the '.env' file to set the environment values, it is
possible to start up the website failed when the user not set up the
'.env' file correctly.

By preserving the default environment values, it could reduce the
change to start up the website failed.

It is weird to load the environment values in the Next.js configuration,
we should load the environment in the 'server.js' when we remove the
Next.js.