Allow using unary ++ and -- operators
Because the unary ++ and -- operators are subject to automatic
semicolon insertion(ASI), differences in white space can change the
semantics of source code.

Since we are using semicolons instead of ASI, it would be fine to use
these operators without any side effect.

By disabling the "no-plusplus" rule to use the unary ++ and --
operators.