Generate key pair on offline
In the practical use case, user will generate the key pair and manage
it by itself, it is improper to generate it from the server-side API,
so we remove the 'keyPairs' API and generate the key pair on the
offline.

Since the 'keyPairs' API is removed, remove 'fetchKeyParis' action,
'keyPairs' reducer and constants as well.