feature: simplify identifier registration
Using the "tic.api.js" API to register the identifier spends lots of
time, because it create multiple channels.

To simplify the registration workflow, we use the "mam.tools.js" to
replace the "tic.api.js", and create a channel to store each revision
of DID document, and use the channel root of the first message to
generated DID.

The DID schema is changed to "did:tangle:{channel-root}_".

The "tic.api.js" package is dropped.